patent_classification_project/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.1.4
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
joblib==0.17.0
nltk==3.5
numpy==1.19.4
pandas==1.1.5
pickles==0.1.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-epo-ops-client==3.1.3
pytz==2020.4
regex==2020.11.13
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
soupsieve==2.1
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.54.1
urllib3==1.26.2
whitenoise==5.2.0
xgboost==1.3.0.post0
